Update fa.json
After using this bot on telegram, I found this translation need some modifications.
Local money in Iran is toman(تومان) so i change all amount to it.
In line 42 the meaning was the opposite of what it should be!
In Iran scooters are known as scooter kids! the scooter that looks like motorcycle is known as motorcycle!
This translation needs more attentions and modifications because persian is  a right to left language! and this translation is very traditional! and seems real robot is speaking so I think this should changes to native and fluent translation.